Update README.md
"Blind Man Shoe's" Shoes for blind men help orassist optically handicapped persons to reach their respective terminus. Itpresents a working conceptual model to provide a clever automated wearabledevice for blind people. The structure is proposed to offer overall solutions whichinclude artificial vision and object detection. Ultrasonic sensors help incalculating the distance of the hurdles lying in the path of blind person. Thisdevice helps in providing the best available track to the user in the form ofvibration. The sensors are integrated with the shoe so that the blind person cannot only detect obstacles in front but also detect the presence of any majorobstacles on the way.